+++ src/components/CardNews/ChipContainer.tsx
@@ -1,16 +1,20 @@
 import { NewsData } from "@/interfaces/NewsData"
 import styled from "@emotion/styled"
 
-export default function ChipContainer({symbol, logo}:Pick<NewsData, 'symbol'|'logo'>) {
+export default function ChipContainer({symbol, logo='ddd'}:Pick<NewsData, 'symbol'|'logo'>) {
   return (
     <ChipsWrapper>
       <Chip>
-        <div>{symbol}</div>
         <div>{logo}</div>
+        <div>{symbol}</div>
       </Chip>
       <Chip>
-        <div>{symbol}</div>
         <div>{logo}</div>
+        <div>{symbol}</div>
+      </Chip>
+      <Chip>
+        <div>{logo}</div>
+        <div>{symbol}</div>
       </Chip>
     </ChipsWrapper>
   )
@@ -20,41 +24,44 @@ const ChipsWrapper = styled.div`
   display: flex;
   align-items: center;
   padding: 0 0 5px;
+  z-index: -1;
 `
 
 const Chip = styled.div`
+box-sizing: border-box;
+height: 28px;
   border-top: 1.5px solid #E5E7EC;
   border-bottom: 1.5px solid #E5E7EC;
   display: inline-flex;
   justify-content: center;
   align-items: center;
-  height: 28px;
-  padding: 8px 0;
+  gap: 5px;
   position: relative;
-  background-color: #fff;
   font-size: 12px;
   margin: 0 20px;
+  div {
+    background-color: #fff;
+    z-index: 9;
+    padding: 3.5px 0;
+  }
   &::before {
-    content: ' ';
+    content: '';
     position: absolute;
-    z-index:-1;
     left: -14px;
     width: 28px;
     height: 28px;
     border-radius: 50%;
-    background-color: red;
     border:1.5px solid #E5E7EC;
   }
   &::after {
-    content: ' ';
+    content: '';
     position: absolute;
-    z-index:-1;
     right: -14px; 
     width: 28px;
     height: 28px;
     border-radius: 50%;
     border: 1.5px solid #E5E7EC;
-    background-color: red;
+
   }
 `
